requirements.txt
clip==0.2.0
datasets==2.16.1
diffusers==0.25.0
Django==5.0.1
matplotlib==3.6.1
numpy==1.23.4
openai==1.7.2
pandas==2.1.4
Pillow==9.3.0
Pillow==10.2.0
PyJWT==2.8.0
Requests==2.31.0
streamlit==1.30.0
streamlit_chat==0.1.1
torch==1.13.0
torchmetrics==1.3.0
torchvision==0.14.0
tqdm==4.66.1
webdataset==0.2.86